pyproject.toml
[tool.poetry]
name = "dbt-llm-tools"
version = "0.2.0"
description = "LLM based tools for dbt projects. Answer data questions, generate documentation and more."
authors = [
    "Pragun Bhutani <1109752+pragunbhutani@users.noreply.github.com>"
]
license = "MIT"
readme = "README.md"
repository = "https://github.com/pragunbhutani/dbt-llm-tools"
keywords = ["dbt", "openai", "llm", "data chatbot", "dbt documentation"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

packages = [
    { include = "dbt_llm_tools" },
]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.16.0"
chromadb = "^0.4.24"
PyYAML = "^6.0.1"
typing-extensions = "^4.10.0"
streamlit = "^1.33.0"
tinydb = "^4.8.0"
boto3 = "^1.37.9"

[tool.poetry.group.dev.dependencies]
pylint = "^3.1.0"
flake8 = "^7.0.0"
black = "^24.3.0"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
flake8-pyproject = "^1.2.3"


[tool.poetry.group.local.dependencies]
notebook = "^7.1.2"


[tool.poetry.group.docs.dependencies]
sphinx-rtd-theme = "^2.0.0"

[tool.flake8]
max-line-length = 120
per-file-ignores = """
    __init__.py: F401
"""

[tool.pylint.'MESSAGES CONTROL']
max-line-length=120
disable = """
    missing-module-docstring,
    too-many-arguments,
    too-few-public-methods,
    broad-exception-raised,
    missing-function-docstring,
    redefined-builtin,
    duplicate-code
"""

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
